# apps/ios/LorealClienteling/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/zones": {
            "get": {
                "operationId": "ZonesController_findAll",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Zones"
                ]
            },
            "post": {
                "operationId": "ZonesController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateZoneDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Zones"
                ]
            }
        },
        "/zones/{id}": {
            "get": {
                "operationId": "ZonesController_findOne",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Zones"
                ]
            },
            "patch": {
                "operationId": "ZonesController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateZoneDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Zones"
                ]
            }
        },
...
        "/stores": {
            "get": {
                "operationId": "StoresController_findAll",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Stores"
                ]
            },
            "post": {
                "operationId": "StoresController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateStoreDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Stores"
                ]
            }
        },
        "/stores/{id}": {
            "get": {
                "operationId": "StoresController_findOne",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Stores"
                ]
            },
            "patch": {
                "operationId": "StoresController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateStoreDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Stores"
                ]
            }
        },
...
        "/customers/{customerId}/recommendations": {
            "get": {
                "operationId": "RecommendationsController_findByCustomer",
                "parameters": [
                    {
                        "name": "customerId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Recommendations"
                ]
            }
        },
        "/recommendations": {
            "post": {
                "operationId": "RecommendationsController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateRecommendationDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Recommendations"
                ]
            }
        },
        "/recommendations/ai": {
            "post": {
                "operationId": "RecommendationsController_requestAi",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AiRecommendationRequestDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Recommendations"
                ]
            }
        },
        "/recommendations/{id}/convert": {
            "patch": {
                "operationId": "RecommendationsController_markConverted",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Recommendations"
                ]
            }
        },
        "/customers/{customerId}/purchases": {
            "get": {
                "operationId": "PurchasesController_findByCustomer",
                "parameters": [
                    {
                        "name": "customerId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Purchases"
                ]
            }
        },
        "/purchases": {
            "post": {
                "operationId": "PurchasesController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreatePurchaseDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Purchases"
                ]
            }
        },
        "/purchases/{id}": {
            "get": {
                "operationId": "PurchasesController_findOne",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Purchases"
                ]
            }
        },
        "/customers/{customerId}/samples": {
            "get": {
                "operationId": "SamplesController_findByCustomer",
                "parameters": [
                    {
                        "name": "customerId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Samples"
                ]
            }
        },
        "/samples": {
            "post": {
                "operationId": "SamplesController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateSampleDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Samples"
                ]
            }
        },
        "/samples/{id}/convert": {
            "patch": {
                "operationId": "SamplesController_markConverted",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Samples"
                ]
            }
        },
...
    },
    "info": {
        "title": "L'Or\u00e9al Clienteling API",
        "description": "API for L'Or\u00e9al beauty advisor clienteling platform",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [
        {
            "url": "http://localhost:3001",
            "description": "Development"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
            "CreateZoneDto": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "example": "CDMX-NOR",
                        "minLength": 1,
                        "maxLength": 50
                    },
                    "displayName": {
                        "type": "string",
                        "example": "CDMX Norte",
                        "minLength": 1,
                        "maxLength": 200
                    },
                    "region": {
                        "type": "string",
                        "example": "Centro",
                        "maxLength": 200
                    }
                },
                "required": [
                    "code",
                    "displayName"
                ]
            },
            "UpdateZoneDto": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "example": "CDMX-NOR",
                        "minLength": 1,
                        "maxLength": 50
                    },
                    "displayName": {
                        "type": "string",
                        "example": "CDMX Norte",
                        "minLength": 1,
                        "maxLength": 200
                    },
                    "region": {
                        "type": "string",
                        "example": "Centro",
                        "maxLength": 200
                    }
                }
            },
...
            "CreateStoreDto": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "example": "LIV-SANTA-FE",
                        "minLength": 1,
                        "maxLength": 50
                    },
                    "displayName": {
                        "type": "string",
                        "example": "Liverpool Santa Fe",
                        "minLength": 1,
                        "maxLength": 200
                    },
                    "chain": {
                        "type": "string",
                        "enum": [
                            "liverpool",
                            "palacio",
                            "owned"
                        ],
                        "example": "liverpool"
                    },
                    "zoneId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "address": {
                        "type": "string",
                        "maxLength": 300
                    },
                    "city": {
                        "type": "string",
                        "example": "Ciudad de M\u00e9xico",
                        "maxLength": 100
                    },
                    "state": {
                        "type": "string",
                        "example": "CDMX",
                        "maxLength": 100
                    }
                },
                "required": [
                    "code",
                    "displayName",
                    "chain"
                ]
            },
            "UpdateStoreDto": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "example": "LIV-SANTA-FE",
                        "minLength": 1,
                        "maxLength": 50
                    },
                    "displayName": {
                        "type": "string",
                        "example": "Liverpool Santa Fe",
                        "minLength": 1,
                        "maxLength": 200
                    },
                    "chain": {
                        "type": "string",
                        "enum": [
                            "liverpool",
                            "palacio",
                            "owned"
                        ],
                        "example": "liverpool"
                    },
                    "zoneId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "address": {
                        "type": "string",
                        "maxLength": 300
                    },
                    "city": {
                        "type": "string",
                        "example": "Ciudad de M\u00e9xico",
                        "maxLength": 100
                    },
                    "state": {
                        "type": "string",
                        "example": "CDMX",
                        "maxLength": 100
                    }
                }
            },
...
            "CreateRecommendationDto": {
                "type": "object",
                "properties": {
                    "customerId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "productId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "source": {
                        "type": "string",
                        "enum": [
                            "manual",
                            "ai_suggested",
                            "replenishment_alert"
                        ],
                        "example": "manual"
                    },
                    "visitReason": {
                        "type": "string",
                        "enum": [
                            "new_purchase",
                            "rebuy",
                            "gift",
                            "concern",
                            "promotion",
                            "browsing"
                        ]
                    },
                    "aiReasoning": {
                        "type": "string"
                    },
                    "notes": {
                        "type": "string",
                        "maxLength": 1000
                    }
                },
                "required": [
                    "customerId",
                    "productId",
                    "source"
                ]
            },
            "AiRecommendationRequestDto": {
                "type": "object",
                "properties": {
                    "customerId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "context": {
                        "type": "string",
                        "maxLength": 2000
                    }
                },
                "required": [
                    "customerId"
                ]
            },
            "PurchaseItemDto": {
                "type": "object",
                "properties": {
                    "productId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "sku": {
                        "type": "string",
                        "example": "SKU-001",
                        "minLength": 1
                    },
                    "quantity": {
                        "type": "number",
                        "example": 2,
                        "minimum": 1
                    },
                    "unitPrice": {
                        "type": "number",
                        "example": 1299,
                        "minimum": 0
                    }
                },
                "required": [
                    "productId",
                    "sku",
                    "quantity",
                    "unitPrice"
                ]
            },
            "CreatePurchaseDto": {
                "type": "object",
                "properties": {
                    "customerId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "source": {
                        "type": "string",
                        "enum": [
                            "pos_integration",
                            "manual",
                            "ecommerce"
                        ],
                        "example": "manual"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PurchaseItemDto"
                        }
                    },
                    "totalAmount": {
                        "type": "number",
                        "example": 2598,
                        "minimum": 0
                    },
                    "posTransactionId": {
                        "type": "string",
                        "example": "POS-12345"
                    }
                },
                "required": [
                    "customerId",
                    "source",
                    "items",
                    "totalAmount"
                ]
            },
            "CreateSampleDto": {
                "type": "object",
                "properties": {
                    "customerId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "productId": {
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "required": [
                    "customerId",
                    "productId"
                ]
            },
...
        }
    }
}